Apply quality filters after imputation by Beagle5.0

Remove markers from the imputed VCF based on adjustable thresholds. For VCFs imputed by Beagle5

postImputeFilter(
  inPath = NULL,
  inName,
  outPath = NULL,
  outName,
  DR2thresh = 0.75,
  HWEthresh = 1e-20,
  MAFthresh = 0.005
)

Arguments

inPath

path to input VCFs

inName

imputed input VCF name

outPath

path for output to be written to

outName

name desired, don't add file suffix, will be *.vcf.gz

DR2thresh

remove if imputation quality score DR2 supplied in the VCF INFO field by Beagle is less than or equal to the threshold

HWEthresh

remove if p-value HWE from vcftools --hardy less than threshold, smaller p-value means more departure from HWE

MAFthresh

remove if minor allele frequency less than threshold

Value

Details

Uses vcftools and R. of minor allee frequency MAFthresh, p-value measuring likelihood of divergence from hardy-weinberg equilibrium ()

NOTICE: This function is part of a family of functions ("imputation_functions") developed as part of the NextGen Cassava Breeding Project genomic selection pipeline. For some examples of their useage:

IITA_2021GS (should be first use of this package)

NRCRI_2021GS

TARI_2020GS

See also