Predict cross means

Function to predict the mean performances of the offspring of crosses. Takes a list of crosses to predict, marker effects, parental allele dosage matrix as input. Predicts potentially over multiple crosses and multiple traits. With predType="BV" predicts the mid-parent of crosses by computing parental GEBV. With predType="TGV" predicts the mean total merit of cross offspring using a Falconer-MacKay Eqn. 14.6 and takes user supplied additive and dominance effects as input. The additive-dominance effects should be partitioned according to the "genotypic" marker codings (see Vitezica et al. 2013. GENETICS).

predCrossMeans(
  CrossesToPredict,
  predType,
  AddEffectList,
  DomEffectList = NULL,
  doseMat,
  ncores = 1,
  ...
)

Arguments

CrossesToPredict
data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.

predType
string, "BV" or "TGV". "BV" predicts cross mean breeding values as the mean GEBV of parents. "TGV" predicts the cross total genetic value. Warning: prediction of meanTGV with F-M Eqn. 14.6 appropriate only using a+d partition not allele sub. + dom. dev.; genotypic NOT classical in terms used by Vitezica et al. 2013. For that reason, predCrossMeans has a "predType" not a "modelType" argument predType="TGV" uses Falconer-MacKay Eqn. 14.6 and takes add and dom effects. predType="BV" input should be allele subst. effs, computes mid-parent GEBV there is no equivalent to predicting the dominance variance for the mean thus the difference from the predCrossVars() function. NOTICE: NOT SAME as predType argument used in predCrossVars, sorry.

AddEffectList
list of ADDITIVE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.

DomEffectList
list of DOMINANCE effect matrices, one matrix per trait, Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.

doseMat
dosage matrix. required only for modelType=="DirDom". Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID

ncores
number of cores, parallelizes across CrossesToPredict, in multi-trait cases, process traits for each family in serial within each worker.

...

Value

tibble, each row contains predictions for a single cross. Columns:

"Trait":

"sireID":

"damID":

"sireGEBV": genomic estimated breeding value (GEBV) of the male parent of the cross

"damGEBV": genomic estimated breeding value (GEBV) of the female parent of the cross

"predOf": "MeanBV" or "MeanTGV"

"predMean": The predicted mean value for the cross

See also