Read database downloads into R

Read database downloads into R

readDBdata(phenotypeFile, metadataFile = NULL)

Arguments

phenotypeFile

CSV file of field trial phenotype data downloaded from e.g. cassavabase

metadataFile

CSV file of field trial metadata downloaded from e.g. cassavabase

Value

merged data.frame phenos + metadata

Details

NOTICE: This function is part of a family of functions ("cassavabase_pheno_pipeline") developed as part of the NextGen Cassava Breeding Project genomic selection pipeline. For some examples of their useage:

IITA_2021GS (should be first use of this package)

NRCRI_2021GS

TARI_2020GS

IITA_2020GS

See also