Run Beagle5 to impute a target VCF with a reference VCF

Impute with Beagle V5.0. Use an "imputation reference panel". Refer to Beagle documentation for meaning of arguments passed.

runBeagle5(
  targetVCF,
  refVCF,
  mapFile,
  outName,
  nthreads,
  maxmem = "500g",
  impute = TRUE,
  ne = 1e+05,
  samplesToExclude = NULL
)

Arguments

targetVCF

passes to Beagle `gt=targetVCF`

refVCF

passes to Beagle `ref=targetVCF`

mapFile

passes to Beagle `map=mapFile`

outName

passes to Beagle `out=outName (don't put file suffix, Beagle adds *.vcf.gz).

nthreads

passes to Beagle `nthreads=nthreads`

maxmem

passes to java `-Xmx<maxmem>`

impute

passes to Beagle `impute=TRUE`

ne

passes to Beagle `ne=ne`

samplesToExclude

Value

Details

NOTICE: This function is part of a family of functions ("imputation_functions") developed as part of the NextGen Cassava Breeding Project genomic selection pipeline. For some examples of their useage:

IITA_2021GS (should be first use of this package)

NRCRI_2021GS

TARI_2020GS

See also