Run genomic predictions

Run GBLUP model using mmer, potentially on multiple traits. Returns genomic BLUPs (GEBV and GETGV). If requested, returns backsolved marker effects (equivalent to ridge regression / SNP-BLUP). Three models are enabled: additive-only ("A"), additive-plus-dominance ("AD") and a directional-dominance model that incorporates a genome-wide homozygosity effect ("DirDom"). Inbreeding effect is included in output GEBV/GETGV predictions *after* backsolving SNP effects. If requested, returns GEBV/GETGV computed for a selection index using selInd=TRUE and supplying SIwts.

runGenomicPredictions(
  modelType,
  selInd,
  SIwts = NULL,
  getMarkEffs = FALSE,
  returnPEV = FALSE,
  blups,
  grms,
  dosages = NULL,
  gid = "GID",
  ncores = 1,
  nBLASthreads = NULL
)

Arguments

modelType

string, "A", "AD", "DirDom". modelType="A": additive-only, GEBVS modelType="AD": the "classic" add-dom model, GEBVS+GEDDs = GETGVs modelType="DirDom": the "genotypic" add-dom model with prop. homozygous fit as a fixed-effect, to estimate a genome-wide inbreeding effect. obtains add-dom effects, computes allele sub effects (\(\alpha = a + d(q-p)\)) incorporates into GEBV and GETGV

selInd

logical, TRUE/FALSE, selection index accuracy estimates, requires input weights via SIwts

SIwts

required if selInd=FALSE, named vector of selection index weights, names match the "Trait" variable in blups

getMarkEffs

T/F return marker effects, backsolved from GBLUP

returnPEV

T/F return PEVs in GBLUP

blups

nested data.frame with list-column "TrainingData" containing BLUPs. Each element of "TrainingData" list, is data.frame with de-regressed BLUPs, BLUPs and weights (WT) for training and test.

grms

list of genomic relation matrices (GRMs, aka kinship matrices). Any genotypes in the GRMs get predicted with, or without phenotypes. Each element is named either A or D. Matrices supplied must match required by A, AD and DirDom models. e.g. grms=list(A=A,D=D).

dosages

dosage matrix. required only for modelType=="DirDom". Also required if getMarkEffs==TRUE. Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID

gid

string variable name used for genotype ID's in e.g. blups (default="GID")

ncores

number of cores

nBLASthreads

number of cores for each worker to use for multi-thread BLAS

Value

tibble, one row, two list columns (basically a named two-element list of lists): gblups[[1]] and genomicPredOut[[1]]. codegblups[[1]]: tibble of predicted GEBV/GETGV, all traits and potentially SELIND genomic BLUPs along the columns. genomicPredOut[[1]] is a tibble that contains some combination of lists-columns:

gblups

varcomps,

fixeffs,

allelesubsnpeff,

addsnpeff,

domstar_snpeff,

domsnpeff

See also