Create a matrix of pairwise recombination frequencies from a genetic map

Compute the pairwise recombination frequencies between all loci from genetic map positions.

genmap2recombfreq(m, nChr)

Arguments

m
vector of centiMorgan-scale genetic positions. names(m) correspond to a SNP_ID. Since m potentially contains all chromosomes, sets recomb. freq. b/t chrom. to 0.5

nChr
number of chromosomes

Value

potentially really large matrix of pairwise recombination frequencies between loci

Details

names(m) must be formatted as "chr"_"id" with "chr" being integer. For example: 2_QTL1 for a locus on chr. 2. May be worth computing in an R session using multi-threaded BLAS.

See also