Predict crosses

Predict potentially for multiple traits, the means, variances and trait-trait covariances in a set ofuser supplied crosses.l If requested, computed the selection index means and variances. Computes the usefulness criteria \(UC_{parent}\) and \(UC_{variety}\) potentially with a user supplied standardized selection intensity value stdSelInt. Output enables easy ranking of potential crosses. This function takes the matrices of snpeffects output (genomicPredOut[[1]]) from the runGenomicPredictions function (when getMarkEffs=TRUE). This is a wrapper function around predCrossVars and predCrossMeans.

predictCrosses(
  modelType,
  stdSelInt = 2.67,
  selInd,
  SIwts = NULL,
  CrossesToPredict,
  snpeffs,
  dosages,
  haploMat,
  recombFreqMat,
  ncores = 1,
  nBLASthreads = NULL
)

Arguments

modelType
string, A, AD or DirDom. A and AD representing model with

selInd
logical, TRUE/FALSE, selection index accuracy estimates, requires input weights via SIwts

CrossesToPredict
data.frame or tibble, col/colnames: sireID, damID. sireID and damID must both be in the haploMat.

snpeffs
the element genomicPredOut[[1]] of the output of runGenomicPredictions.

dosages
dosage matrix. required only for modelType=="DirDom". Assumes SNPs coded 0, 1, 2. Nind rows x Nsnp cols, numeric matrix, with rownames and colnames to indicate SNP/ind ID

haploMat
matrix of phased haplotypes, 2 rows per sample, cols = loci, 0,1, rownames assumed to contain GIDs with a suffix, separated by "_" to distinguish haplotypes

recombFreqMat
a square symmetric matrix with values = (1-2*c1), where c1=matrix of expected recomb. frequencies. The choice to do 1-2c1 outside the function was made for computation efficiency; every operation on a big matrix takes time.

ncores
number of cores

nBLASthreads
number of cores for each worker to use for multi-thread BLAS

Value

tibble, one row, two list columns (basically a named two-element list of lists): tidyPreds[[1]] and rawPreds[[1]]. codetidyPreds[[1]]: tidy output, fewer details. sireID, damID, Nsegsnps, predOf,Trait, predMean, predVar, predSD, predUsefulnesstibble of predicted GEBV/GETGV, all traits and potentially SELIND genomic BLUPs along the columns. rawPreds[[1]]: more detailed output, list of 2 ("predMeans" tibble and "predVars" tibble).

See also